+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.42.2
 - Improve download-core command output
+- Ignore missing requests to client responses
 
 ## 0.42.1 - June 11, 2021
 - YardMap#change sets new directory (#445)